CBild with import and export
Changed CBild to be initialized with the key that is supposed to be used
for encoding.
CBild does NOT inherit from NBild but has methods to import an NBild and
export arbitrary NBild objects to file, so we can export the original as well as
the encoded image.
Encode and decode should now work with whatever file has been imported
and use the key the object has been initialised with.